Adding EPIC, CCDMinimizer, and minimization of pure pairwise energy
Adding the EPIC (Energy as Polynomials in Internal Coordinates)
algorithm, turned on using the useEPIC flag.  Also providing a faster
and more accurate minimizer, the CCDMinimizer (turned on by the useCCD
flag, which is on by default).  Previous versions of OSPREY minimized
the full energy of a rotamer pair when trying to get the minimum
pairwise energy, which is not provable though it results in higher
bounds than minimizing the pairwise energy alone; default behavior is
now to minimize the pairwise energy alone (the old heuristic can be
enabled with minimizePairwise=false).  Finally, the makeStericShell
top-level command has been added.